# Replace name_to_labels(s).last with faster extract_tld