Following https://github.com/weppos/publicsuffix-ruby/pull/128 I looked into `PublicSuffix::Domain.name_to_labels`'s performance.

The thing is, it's almost always used to get the TLD, so I benchmark several implementations:

```ruby
require 'benchmark/ips'

DOMAIN = 'accident-prevention.aero'
DOT = '.'
Benchmark.ips do |x|
  x.report('string.rindex + string.slice') {
    DOMAIN[(DOMAIN.rindex('.')+1..-1)]
  }
  x.report('string.rpartition') {
    DOMAIN.rpartition(DOT)[2]
  }
  x.report('split.last') {
    DOMAIN.split('.').last
  }
  x.report('string.scan') {
    DOMAIN.scan(/[^\.]+\z/).first
  }
  x.report('regexp.match') {
    /[^\.]+\z/.match(DOMAIN)[0]
  }
  x.report('string.slice') {
    DOMAIN.slice(/[^\.]+\z/)
  }
end
```

```
string.rindex + string.slice       1.710M (± 6.9%) i/s -      8.539M in   5.020714s
           string.rpartition       2.214M (± 7.7%) i/s -     11.095M in   5.044656s
                  split.last       1.716M (± 4.2%) i/s -      8.638M in   5.044117s
                 string.scan      86.526k (± 7.0%) i/s -    434.673k in   5.053164s
                regexp.match     157.309k (± 8.2%) i/s -    794.035k in   5.084998s
                string.slice     191.887k (± 4.8%) i/s -    972.379k in   5.080122s
```

So using `string.rpartition('.')[2]` is 30% faster than `string.split('.').last`.

It probably won't make a huge difference, but it might shave a few ms off the `reindex!`.

@weppos thoughts ?