@weppos 2017-01-22
See https://github.com/weppos/publicsuffix-ruby/pull/130#issuecomment-274317007